Rename parameters file for AWS parameters
