+++ integration-tests/lib/awsParameters.ts
@@ -1,4 +1,5 @@
 import { getEnvVariable } from './zendeskParameters'
+
 export const initiateDataRequestLambdalogGroupName =
   '/aws/lambda/ticf-integration-initiate-data-request'
